+++ services/messaging/src/main/java/com/github/punchat/messaging/domain/member/Invitation.java
