+++ services/messaging/src/main/java/com/github/punchat/messaging/domain/message/MessageController.java
@@ -1,0 +1,12 @@
+package com.github.punchat.messaging.domain.message;
+
+import org.springframework.web.bind.annotation.RestController;
+
+@RestController
+public class MessageController {
+    private final MessageService service;
+
+    public MessageController(MessageService service) {
+        this.service = service;
+    }
+}
