+++ services/messaging/src/main/java/com/github/punchat/messaging/domain/message/MessageService.java
@@ -1,0 +1,4 @@
+package com.github.punchat.messaging.domain.message;
+
+public interface MessageService {
+}
