+++ services/messaging/src/main/java/com/github/punchat/messaging/domain/role/Permission.java
@@ -1,4 +1,4 @@
-package com.github.punchat.messaging.domain.member;
+package com.github.punchat.messaging.domain.role;
 
 public enum Permission {
     CAN_INVITE_USERS, CAN_EXCLUDE_USERS, CAN_WRITE_MESSAGES
