+++ services/messaging/src/main/resources/hibernate.cfg.xml
@@ -13,6 +13,6 @@
         <mapping class="com.github.punchat.messaging.domain.message.DirectMessage"/>
         <mapping class="com.github.punchat.messaging.domain.message.BroadcastMessage"/>
         <mapping class="com.github.punchat.messaging.domain.member.Member"/>
-        <mapping class="com.github.punchat.messaging.domain.member.Role"/>
+        <mapping class="com.github.punchat.messaging.domain.role.Role"/>
     </session-factory>
 </hibernate-configuration>
